doc_to_questions/requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
backcall==0.1.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.0
Markdown==3.1.1
nltk==3.4.4
numpy==1.17.0
pandas==0.25.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textblob==0.15.3
traitlets==4.3.2
Unidecode==1.1.1
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
